requirements.txt
python-telegram-bot==21.5
google-api-python-client==2.143.0
google-auth-oauthlib==1.2.1
google-auth==2.34.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
